flytt validering av årsak til valideringsfil
